Services/SpatialMemory/requirements.txt
fastapi>=0.116.0
uvicorn>=0.35.0
pydantic>=2.11.0
numpy>=2.1.0
Pillow>=10.4.0
requests>=2.32.0
